Saved pdf's to separate directory
Minor changes to README.md. I also made sure that all the pdf's were saved to a directory outside of the source. This allows people to view the results without installing Cadabra and friends. It also avoids the problem of trashing source/eqtns.pdf in the event that something went wrong during Cadabra's execution.